# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud HTTP API
  version: 1.0.0
  description: |
    Contrato público de /api/v1. Las rutas /api equivalentes siguen disponibles
    durante la transición y responden con Deprecation y Link al sucesor.
    Las escrituras autenticadas por sesión requieren X-XSRF-TOKEN. El servidor
    acepta X-Correlation-Id y devuelve el identificador efectivo en ese header.
    GET /api/v1/tacos?recent conserva la respuesta de lista del cliente antiguo.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
  /api/v1/ingredients:
    get:
      tags: [Ingredients]
      summary: Listar ingredientes públicos
      security: []
      responses:
        '200': { $ref: '#/components/responses/IngredientList' }
    post:
      tags: [Ingredients]
      summary: Crear ingrediente del catálogo
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/IngredientRequest' }
      responses:
        '201': { $ref: '#/components/responses/IngredientCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ingredients/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Ingredients]
      summary: Consultar ingrediente
      security: []
      responses:
        '200': { $ref: '#/components/responses/Ingredient' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Ingredients]
      summary: Reemplazar ingrediente existente
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/IngredientRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ingredient' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Ingredients]
      summary: Eliminar ingrediente existente
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Crear una orden
      description: |
        Idempotency-Key es opcional para compatibilidad. Con la misma clave y
        el mismo contenido devuelve el mismo orderId y HTTP 201. Otra carga
        con esa clave responde 409. Un claim pendiente también puede responder 409.
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/OrderCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/orders/fromEmail:
    post:
      tags: [Orders]
      summary: Crear una orden desde datos de correo convertidos
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/EmailOrder' }
      responses:
        '201': { $ref: '#/components/responses/OrderCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
  /api/v1/orders/{orderId}:
    parameters:
      - { name: orderId, in: path, required: true, schema: { type: string } }
    put:
      tags: [Orders]
      summary: Reemplazar datos de entrega de una orden editable
      x-required-role: USER o ADMIN propietario
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderReplaceRequest' }
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    patch:
      tags: [Orders]
      summary: Modificar campos de entrega permitidos
      x-required-role: USER o ADMIN propietario
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderPatchRequest' }
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
    delete:
      tags: [Orders]
      summary: Eliminar una orden en estado permitido
      x-required-role: USER o ADMIN propietario
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/orders/{orderId}/status:
    parameters:
      - { name: orderId, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Orders, Admin]
      summary: Cambiar estado de una orden con versión esperada
      x-required-role: ADMIN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/orders/{orderId}/cancel:
    parameters:
      - { name: orderId, in: path, required: true, schema: { type: string } }
    post:
      tags: [Orders]
      summary: Cancelar una orden propia con versión esperada
      x-required-role: USER
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderCancelRequest' }
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/orders/me:
    get:
      tags: [Orders]
      summary: Listar órdenes propias
      x-required-role: USER
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, default: 10 } }
      responses:
        '200': { $ref: '#/components/responses/OrderHistoryPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/orders/me/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      summary: Consultar detalle de una orden propia
      x-required-role: USER
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/orders:
    get:
      tags: [Orders, Admin]
      summary: Auditar órdenes con filtros
      x-required-role: ADMIN
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, default: 10 } }
        - { name: userId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
      responses:
        '200': { $ref: '#/components/responses/OrderHistoryPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders, Admin]
      summary: Consultar detalle de una orden para auditoría
      x-required-role: ADMIN
      responses:
        '200': { $ref: '#/components/responses/Order' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/kitchen/orders/{orderId}/status:
    parameters:
      - { name: orderId, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Kitchen]
      summary: Avanzar una orden asignada a cocina
      x-required-role: KITCHEN
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody: { $ref: '#/components/requestBodies/OrderStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/KitchenOrder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación Basic o sesión autenticada de la UI.
  headers:
    CorrelationId:
      description: Identificador de correlación efectivo para esta petición.
      schema: { type: string, maxLength: 100 }
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: { type: string, pattern: '^[A-Za-z0-9._:-]{1,100}$' }
      description: El servidor genera un ID si el recibido no es válido.
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9_-]+$' }
...
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
      description: Requerido para escrituras autenticadas por sesión. Se obtiene de la cookie XSRF-TOKEN.
  requestBodies:
    IngredientRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/IngredientRequest' }
          example: { id: WRAP, name: Flour tortilla, type: WRAP, dietaryTags: [VEGETARIAN], allergens: [GLUTEN], spiceLevel: NONE }
...
    OrderCreateRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderCreateRequest' }
          example:
            deliveryName: Alice
            deliveryStreet: Street 1
            deliveryCity: Austin
            deliveryState: TX
            deliveryZip: '78701'
            paymentMethodId: PAYMENT-ID
            items:
              - taco: { name: Garden taco, ingredientIds: [WRAP, SLSA] }
                quantity: 2
    EmailOrder:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EmailOrder' }
...
    OrderPatchRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderPatchRequest' }
          example: { deliveryZip: '78702' }
    OrderReplaceRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderReplaceRequest' }
    OrderStatusRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderStatusRequest' }
          example: { status: ACCEPTED, expectedVersion: 1, reason: Accepted by staff }
    OrderCancelRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderCancelRequest' }
          example: { expectedVersion: 1, reason: Customer request }
...
  responses:
    NoContent:
      description: Operación completada sin cuerpo.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
    IngredientList:
      description: Ingredientes públicos.
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/IngredientResponse' }
    Ingredient:
      description: Ingrediente público.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/IngredientResponse' }
    IngredientCreated:
      description: Ingrediente creado.
      headers:
        Location: { description: URI del ingrediente, schema: { type: string, format: uri } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/IngredientResponse' }
...
    Order:
      description: Orden pública sin entidad User ni token de pago.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderResponse' }
    OrderCreated:
      description: Orden creada o respuesta conservada para el mismo Idempotency-Key.
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderResponse' }
          example:
            id: ORDER-ID
            status: CREATED
            userId: USER-ID
            paymentBrand: VISA
            paymentLast4: '0002'
            currency: USD
            subtotal: 3.00
            couponApplied: false
            discount: 0.00
            total: 3.00
            items: []
...
    OrderHistoryPage:
      description: Resúmenes paginados.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OrderHistoryPageResponse' }
...
    KitchenOrder:
      description: Orden asignada a cocina.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/KitchenOrderResponse' }
...
    BadRequest:
      description: Entrada inválida.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
          example: { type: 'urn:tacocloud:problem:malformed-request', title: Bad request, status: 400, detail: The request cannot be processed., instance: /api/v1/orders, code: MALFORMED_REQUEST, violations: [] }
    Unauthorized:
      description: Autenticación requerida. Spring Security puede devolver un 401 sin ApiProblem.
    Forbidden:
      description: Operación prohibida. Spring Security puede devolver un 403 sin ApiProblem.
    NotFound:
      description: Recurso inexistente.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Conflict:
      description: Conflicto de estado, versión o Idempotency-Key.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Unprocessable:
      description: Regla de negocio o diseño inválido.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
  schemas:
    IngredientType:
      type: string
      enum: [WRAP, BOWL, PROTEIN, VEGGIES, CHEESE, SAUCE]
    DietaryTag:
      type: string
      enum: [VEGAN, VEGETARIAN, GLUTEN_FREE]
    Allergen:
      type: string
      enum: [GLUTEN, MILK, EGG, SOY, PEANUTS, TREE_NUTS, FISH, SHELLFISH, SESAME]
    SpiceLevel:
      type: string
      enum: [UNKNOWN, NONE, MILD, MEDIUM, HOT, EXTREME]
    OrderStatus:
      type: string
      enum: [CREATED, ACCEPTED, PLACED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
    IngredientRequest:
      type: object
      additionalProperties: false
      required: [id, name, type]
      properties:
        id: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
        name: { type: string, maxLength: 100, minLength: 1 }
        type: { $ref: '#/components/schemas/IngredientType' }
        dietaryTags:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/DietaryTag' }
        allergens:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Allergen' }
        spiceLevel: { $ref: '#/components/schemas/SpiceLevel' }
    IngredientResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        name: { type: string }
        type: { $ref: '#/components/schemas/IngredientType' }
        unitPrice: { type: number, format: decimal }
        available: { type: boolean }
        dietaryTags:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/DietaryTag' }
        allergens:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Allergen' }
        spiceLevel: { $ref: '#/components/schemas/SpiceLevel' }
...
    OrderCreateItem:
      type: object
      additionalProperties: false
      required: [taco, quantity]
      properties:
        taco: { $ref: '#/components/schemas/OrderCreateTaco' }
        quantity: { type: integer, minimum: 1 }
    OrderCreateRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, paymentMethodId, items]
      properties:
        deliveryName: { type: string, minLength: 1, maxLength: 100 }
        deliveryStreet: { type: string, minLength: 1, maxLength: 150 }
        deliveryCity: { type: string, minLength: 1, maxLength: 100 }
        deliveryState: { type: string, minLength: 1, maxLength: 50 }
        deliveryZip: { type: string, pattern: '^[A-Za-z0-9 -]{3,20}$' }
        paymentMethodId: { type: string, maxLength: 64 }
        couponCode: { type: string, maxLength: 64, pattern: '^[A-Za-z0-9_-]+$' }
        items:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: '#/components/schemas/OrderCreateItem' }
...
    OrderResponseItem:
      type: object
      additionalProperties: false
      properties:
        taco: { $ref: '#/components/schemas/OrderResponseTaco' }
        quantity: { type: integer, minimum: 1 }
        unitPriceAtPurchase: { type: number, format: decimal }
        subtotal: { type: number, format: decimal }
    OrderStatusChange:
      type: object
      additionalProperties: false
      properties:
        from: { $ref: '#/components/schemas/OrderStatus' }
        to: { $ref: '#/components/schemas/OrderStatus' }
        actorId: { type: string }
        actorRole: { type: string }
        changedAt: { type: string, format: date-time }
        origin: { type: string }
        reason: { type: string }
    OrderResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        version: { type: integer, format: int64 }
        statusHistory:
          type: array
          items: { $ref: '#/components/schemas/OrderStatusChange' }
        userId: { type: string }
        deliveryName: { type: string }
        deliveryStreet: { type: string }
        deliveryCity: { type: string }
        deliveryState: { type: string }
        deliveryZip: { type: string }
        paymentBrand: { type: string }
        paymentLast4: { type: string }
        currency: { type: string }
        subtotal: { type: number, format: decimal }
        couponApplied: { type: boolean }
        discount: { type: number, format: decimal }
        total: { type: number, format: decimal }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderResponseItem' }
    EmailTaco:
      type: object
      additionalProperties: false
      required: [name, ingredients]
      properties:
        name: { type: string, minLength: 5, maxLength: 100 }
        ingredients:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: string, maxLength: 64 }
    EmailOrder:
      type: object
      additionalProperties: false
      required: [email, tacos]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        tacos:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: '#/components/schemas/EmailTaco' }
...
    OrderPatchRequest:
      type: object
      additionalProperties: false
      properties:
        deliveryName: { type: string, maxLength: 100 }
        deliveryStreet: { type: string, maxLength: 150 }
        deliveryCity: { type: string, maxLength: 100 }
        deliveryState: { type: string, maxLength: 50 }
        deliveryZip: { type: string, maxLength: 20 }
    OrderReplaceRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip]
      properties:
        deliveryName: { type: string, minLength: 1, maxLength: 100 }
        deliveryStreet: { type: string, minLength: 1, maxLength: 150 }
        deliveryCity: { type: string, minLength: 1, maxLength: 100 }
        deliveryState: { type: string, minLength: 1, maxLength: 50 }
        deliveryZip: { type: string, minLength: 1, maxLength: 20 }
    OrderStatusRequest:
      type: object
      additionalProperties: false
      required: [status, expectedVersion, reason]
      properties:
        status: { $ref: '#/components/schemas/OrderStatus' }
        expectedVersion: { type: integer, format: int64 }
        reason: { type: string, minLength: 1, maxLength: 200 }
    OrderCancelRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, reason]
      properties:
        expectedVersion: { type: integer, format: int64 }
        reason: { type: string, minLength: 1, maxLength: 200 }
    OrderSummaryResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        total: { type: number, format: decimal }
        currency: { type: string }
    OrderHistoryPageResponse:
      type: object
      additionalProperties: false
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/OrderSummaryResponse' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer, format: int64 }
...
    KitchenOrderResponse:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        version: { type: integer, format: int64 }
        stationId: { type: string }
        cookId: { type: string }
        estimatedPrepMinutes: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/KitchenOrderItem' }
...
    ApiProblemViolation:
      type: object
      additionalProperties: false
      properties:
        field: { type: string }
        message: { type: string }
        code: { type: string, nullable: true }
    ApiProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        violations:
          type: array
          items: { $ref: '#/components/schemas/ApiProblemViolation' }